+++ app/components/Views/confirmations/components/activity/transaction-details-summary/deposit-summary-line.test.tsx
@@ -1,23 +1,18 @@
 import React from 'react';
-import { fireEvent } from '@testing-library/react-native';
 import {
   TransactionMeta,
   TransactionType,
 } from '@metamask/transaction-controller';
-import { Hex } from '@metamask/utils';
 import renderWithProvider from '../../../../../../util/test/renderWithProvider';
 import { strings } from '../../../../../../../locales/i18n';
 import { useMultichainBlockExplorerTxUrl } from '../../../../../UI/Bridge/hooks/useMultichainBlockExplorerTxUrl';
-import Routes from '../../../../../../constants/navigation/Routes';
 import { useNetworkName } from '../../../hooks/useNetworkName';
 import { useTokenWithBalance } from '../../../hooks/tokens/useTokenWithBalance';
 import { selectBridgeHistoryForAccount } from '../../../../../../selectors/bridgeStatusController';
 import { useBridgeTxHistoryData } from '../../../../../../util/bridge/hooks/useBridgeTxHistoryData';
 import { useTokenAmount } from '../../../hooks/useTokenAmount';
 import { useTransactionDetails } from '../../../hooks/activity/useTransactionDetails';
-import { RelayDepositSummaryLine } from './relay-deposit-summary-line';
-
-const mockNavigate = jest.fn();
+import { DepositSummaryLine } from './deposit-summary-line';
 
 jest.mock('../../../../../UI/Bridge/hooks/useMultichainBlockExplorerTxUrl');
 jest.mock('../../../hooks/useNetworkName');
@@ -30,13 +25,13 @@ jest.mock('../../../hooks/activity/useTransactionDetails');
 jest.mock('@react-navigation/native', () => ({
   ...jest.requireActual('@react-navigation/native'),
   useNavigation: () => ({
-    navigate: mockNavigate,
+    navigate: jest.fn(),
   }),
 }));
 
 function render(parentType: TransactionType = TransactionType.perpsDeposit) {
   return renderWithProvider(
-    <RelayDepositSummaryLine
+    <DepositSummaryLine
       transactionMeta={
         {
           id: 'relay-id',
@@ -51,12 +46,12 @@ function render(parentType: TransactionType = TransactionType.perpsDeposit) {
           id: 'parent-id',
           chainId: '0x1',
           type: parentType,
+          metamaskPay: {
+            tokenAddress: '0x123',
+            chainId: '0x1',
+          },
         } as unknown as TransactionMeta
       }
-      payTokenInfo={{
-        tokenAddress: '0x123' as Hex,
-        chainId: '0x1' as Hex,
-      }}
     />,
     {
       state: {
@@ -70,7 +65,7 @@ function render(parentType: TransactionType = TransactionType.perpsDeposit) {
   );
 }
 
-describe('RelayDepositSummaryLine', () => {
+describe('DepositSummaryLine', () => {
   const useMultichainBlockExplorerTxUrlMock = jest.mocked(
     useMultichainBlockExplorerTxUrl,
   );
@@ -129,17 +124,18 @@ describe('RelayDepositSummaryLine', () => {
     ).toBeDefined();
   });
 
-  it('navigates to block explorer when button is pressed', () => {
-    const { getByTestId } = render();
+  it('renders loading title when token or network unavailable', () => {
+    useTokenWithBalanceMock.mockReturnValue(
+      {} as ReturnType<typeof useTokenWithBalance>,
+    );
+    useNetworkNameMock.mockReturnValue(undefined);
 
-    fireEvent.press(getByTestId('block-explorer-button'));
+    const { getByText } = render();
 
-    expect(mockNavigate).toHaveBeenCalledWith(Routes.WEBVIEW.MAIN, {
-      screen: Routes.WEBVIEW.SIMPLE,
-      params: {
-        url: 'https://explorer.example',
-        title: 'Explorer',
-      },
-    });
+    expect(
+      getByText(
+        strings('transaction_details.summary_title.bridge_send_loading'),
+      ),
+    ).toBeDefined();
   });
 });
