+++ app/components/Views/confirmations/components/activity/transaction-details-summary/transaction-summary-line.tsx
@@ -5,12 +5,13 @@ import {
   TransactionStatus,
 } from '@metamask/transaction-controller';
 import { Hex } from '@metamask/utils';
+import I18n from '../../../../../../../locales/i18n';
 import { IconName } from '../../../../../../component-library/components/Icons/Icon';
 import Routes from '../../../../../../constants/navigation/Routes';
+import { getIntlDateTimeFormatter } from '../../../../../../util/intl';
 import { useMultichainBlockExplorerTxUrl } from '../../../../../UI/Bridge/hooks/useMultichainBlockExplorerTxUrl';
 import { ProgressListItem } from '../../progress-list';
 import { Severity } from '../../status-icon';
-import { getDateString } from './utils';
 
 export function TransactionSummaryLine({
   chainId,
@@ -71,6 +72,24 @@ export function TransactionSummaryLine({
       onButtonPress={explorerTxUrl ? handleExplorerClick : undefined}
     />
   );
+}
+
+function getDateString(timestamp: number): string {
+  const date = new Date(timestamp);
+
+  const timeString = getIntlDateTimeFormatter(I18n.locale, {
+    hour: 'numeric',
+    minute: 'numeric',
+    hour12: true,
+  }).format(date);
+
+  const month = getIntlDateTimeFormatter(I18n.locale, {
+    month: 'short',
+  }).format(date);
+
+  const dateString = `${month} ${date.getDate()}, ${date.getFullYear()}`;
+
+  return `${timeString} • ${dateString}`;
 }
 
 function getSeverity(status: TransactionStatus): Severity {
