--- app/components/Views/confirmations/components/activity/transaction-details-summary/types.ts
@@ -1,6 +1,0 @@
-import { Hex } from '@metamask/utils';
-
-export interface PayTokenInfo {
-  tokenAddress: Hex | undefined;
-  chainId: Hex | undefined;
-}
