--- app/components/Views/confirmations/components/activity/transaction-details-summary/utils.ts
@@ -1,20 +1,0 @@
-import I18n from '../../../../../../../locales/i18n';
-import { getIntlDateTimeFormatter } from '../../../../../../util/intl';
-
-export function getDateString(timestamp: number): string {
-  const date = new Date(timestamp);
-
-  const timeString = getIntlDateTimeFormatter(I18n.locale, {
-    hour: 'numeric',
-    minute: 'numeric',
-    hour12: true,
-  }).format(date);
-
-  const month = getIntlDateTimeFormatter(I18n.locale, {
-    month: 'short',
-  }).format(date);
-
-  const dateString = `${month} ${date.getDate()}, ${date.getFullYear()}`;
-
-  return `${timeString} • ${dateString}`;
-}
