+++ app/components/Views/confirmations/components/progress-list/index.ts
@@ -1,1 +1,2 @@
 export * from './progress-list';
+export * from './progress-list-item';
